chore: move test helpers to separate files
