+++ EssentialFeed/EssentialFeediOSTests/Helpers/FeedViewControllerTests+FakeRefreshControl.swift
@@ -1,0 +1,22 @@
+//
+//  FeedViewControllerTests+FakeRefreshControl.swift
+//  EssentialFeediOSTests
+//
+//  Created by Nikhil Menon on 2/13/24.
+//
+
+import UIKit
+
+class FakeRefreshContol: UIRefreshControl {
+    private var _isRefreshing: Bool = false
+    
+    override var isRefreshing: Bool { _isRefreshing }
+    
+    override func beginRefreshing() {
+        _isRefreshing = true
+    }
+    
+    override func endRefreshing() {
+        _isRefreshing = false
+    }
+}
