+++ EssentialFeed/EssentialFeediOSTests/Helpers/UIImage+TestHelpers.swift
@@ -1,0 +1,21 @@
+//
+//  UIImage+TestHelpers.swift
+//  EssentialFeediOSTests
+//
+//  Created by Nikhil Menon on 2/13/24.
+//
+
+import UIKit
+
+extension UIImage {
+    static func make(withColor color: UIColor) -> UIImage {
+        let rect = CGRect(x: 0, y: 0, width: 1, height: 1)
+        let format = UIGraphicsImageRendererFormat()
+        format.scale = 1
+        
+        return UIGraphicsImageRenderer(size: rect.size, format: format).image { context in
+            color.setFill()
+            context.fill(rect)
+        }
+    }
+}
