+++ EssentialFeed/EssentialFeediOSTests/Helpers/UIRefreshControl+TestHelpers.swift
@@ -1,0 +1,18 @@
+//
+//  UIRefreshControl+TestHelpers.swift
+//  EssentialFeediOSTests
+//
+//  Created by Nikhil Menon on 2/13/24.
+//
+
+import UIKit
+
+extension UIRefreshControl {
+    func simulatePullToRefresh() {
+        allTargets.forEach { target in
+            actions(forTarget: target, forControlEvent: .valueChanged)?.forEach { action in
+                (target as NSObject).perform(Selector(action))
+            }
+        }
+    }
+}
